Move JSON:API error handing to a dedicated module
